requirements.txt
undetected_chromedriver==3.5.5
setuptools==80.9.0
pandas==2.3.3
langchain==1.2.4
langchain-community==0.4.1
langchain-openai==1.1.7
python-dotenv==1.2.1
langgraph==1.2.7
pyyaml==6.0.3
